requirements.txt
pandas
sqlalchemy
pymysql
python-dotenv
requests
pdfplumber
shortuuid
openpyxl
flask
flask_mail
fuzzywuzzy
levenshtein